1) Renamed anonymize and original DICOM folders to include PatientName 2) Zip automatically anonymize and original DICOM folders (redmine ticket https://redmine.cbrain.mcgill.ca/issues/8073)
